fix: memory pool structure fixed

+++ MemoryPool/CNonObject.h
@@ -9,7 +9,7 @@ public:
     ~CNonObject() {};
 
 public:
-    float X;
-    float Y;
-    float Z;
+    double X;
+    double Y;
+    double Z;
 };
